deleteExercise unit test - Tinashe Chamisa
