+++ nest-back/test/Workout/UnitTests/updateExercise.spec.ts
@@ -16,65 +16,21 @@ beforeEach(() => {
     ctx = (mockCtx as unknown) as Context
 })
 
-test('Valid exercise passed in with tags, should receive successful message', async () => {
-    const Exercise = {
-        exercise:uuidExercise,
-        title:"test",
-        description:"test",
-        repRange:"test",
-        sets:4,
-        Posedescription:"test",
-        restPeriod:2,
-        tags: [{
-            label:'test',
-            textColour:'test',
-            backgroundColour:'test'
-        }],
-        duratime:2,
-        workouts: null
-    }
-    mockCtx.prisma.exercise.update.mockResolvedValue(Exercise)
-
-    const response=await workoutService.updateExercise(uuidExercise,'test','test','test',4,'test',2,[{label:'test',textColour:'test',backgroundColour:'test'}],2,ctx)
-    console.log(response);
-    expect(response).toBe("Exercise updated.");
-})
-
-test('Valid exercise passed in without tags, should receive successful message', async () => {
-    const Exercise = {
-        exercise:uuidExercise,
-        title:"test",
-        description:"test",
-        repRange:"test",
-        sets:4,
-        Posedescription:"test",
-        restPeriod:2,
-        tags: [],
-        duratime:2,
-        workouts: null
-    }
-    mockCtx.prisma.exercise.update.mockResolvedValue(Exercise)
-
-    const response=await workoutService.updateExercise(uuidExercise,'test','test','test',4,'test',2,[{label:'test',textColour:'test',backgroundColour:'test'}],2,ctx)
-
-    expect(response).toBe("Exercise updated.");
-})
-
 test('Null exercise passed in, should throw PreconditionFailedException', async () => {
-    let tag;
-    mockCtx.prisma.tag.create.mockResolvedValue(tag)
+    let exercise;
+    mockCtx.prisma.tag.update.mockResolvedValue(exercise)
     await expect(workoutService.updateExercise('','','','',0,'',0,[],0,ctx)).rejects.toThrow("Invalid exercise object passed in.")
 })
 
 test('Incomplete exercise passed in, should throw PreconditionFailedException', async () => {
-    let tag;
-    mockCtx.prisma.tag.create.mockResolvedValue(tag)
+    let exercise;
+    mockCtx.prisma.tag.update.mockResolvedValue(exercise)
     await expect(workoutService.updateExercise('test','','test','',0,'test',0,[],0,ctx)).rejects.toThrow("Invalid exercise object passed in.")
 })
 
 test('Nonexistent exercise, should throw NotFoundException', async () => {
-    let tag;
-    mockCtx.prisma.tag.create.mockResolvedValue(tag)
+    let exercise;
+    mockCtx.prisma.tag.update.mockResolvedValue(exercise)
     await expect(workoutService.updateExercise('test','test','test','test',0,'test',0,[],0,ctx)).rejects.toThrow("Exercise with provided ID does not exist.")
 })
